Add more entropy into post signature to ensure uniqueness
Per our discussion yesterday, post signatures are now created with
* the title
* the content
* the created timestamp
* the last updated timestamp

There's an implication here to call out: when you create, the nextJS must also
pass in the lastUpdatedAt field.

This simply provides an air of completeness, that is to say: anything the dApp
submitted has been signed...you cannot tamper with ANY of the post fields.