Refactor how we view and manage posts in NextJS app
In an effort to make our lives easier, this change aims to make it so that we
have one Post component that is configurable to render different sub-components
in different contexts, but provide consistent display & functional experiences.

* Both `/feed` and `/feed/viewPost` use the same Post component
* The Post component expects a post ID and then takes care of retrieving post
data for itself and it's subcomponents using the `getPost` contract method
* All edit & delete logic have been pushed in to the Post component
  * The PostHeader has the buttons to trigger edit/delete modals
  * Edit/Delete modals were moved into their own components
  * There is an edit/delete context provider to facilitate communication between
    the buttons and the modals.
* The Post component has a display context provider which parent nodes can
configure to show a compact or expanded view of Post subcomponents. This is only
being actively used for how we dispaly the creator username & address, but can
easily be expanded on.
* On `/feed`, instead of using the `getCommunityPosts` to render the posts, we
instead use it for the post IDs - we should change the contract in a future
iteration.

This is a first iteration refactor but I think should make our lives easier.
It's a good stopping point, but I think one more pass would useful in order to
keep the edit/delete logic even closer to their respective components and
actually drop the context providers...and if not drop them completely, at least
drastically simplify them.